move code to challenges app
