+++ apps/challenges/models.py
@@ -13,6 +13,7 @@ from base.models import (
 )
 from base.utils import RandomFileName
 from participants.models import ParticipantTeam
+from hosts.models import ChallengeHost
 
 
 class Challenge(TimeStampedModel):
@@ -331,3 +332,30 @@ class StarChallenge(TimeStampedModel):
     class Meta:
         app_label = "challenges"
         db_table = "starred_challenge"
+
+
+class UserInvitation(TimeStampedModel):
+    """
+    Model to store invitation status
+    """
+
+    ACCEPTED = "accepted"
+    PENDING = "pending"
+
+    STATUS_OPTIONS = ((ACCEPTED, ACCEPTED), (PENDING, PENDING))
+    email = models.EmailField(max_length=200)
+    invitation_key = models.CharField(max_length=200)
+    status = models.CharField(
+        max_length=30, choices=STATUS_OPTIONS, db_index=True
+    )
+    challenge = models.ForeignKey(Challenge, related_name="challenge")
+    user = models.ForeignKey(User)
+    invited_by = models.ForeignKey(ChallengeHost)
+
+    class Meta:
+        app_label = "challenges"
+        db_table = "invite_user_to_challenge"
+
+    def __str__(self):
+        """Returns the email of the user"""
+        return self.email
