+++ apps/challenges/serializers.py
@@ -1,5 +1,6 @@
 from rest_framework import serializers
 
+from accounts.serializers import UserDetailsSerializer
 from hosts.serializers import ChallengeHostTeamSerializer
 
 from .models import (
@@ -10,6 +11,7 @@ from .models import (
     DatasetSplit,
     Leaderboard,
     StarChallenge,
+    UserInvitation,
 )
 
 
@@ -269,3 +271,37 @@ class StarChallengeSerializer(serializers.ModelSerializer):
             challenge=obj.challenge, is_starred=True
         ).count()
         return count
+
+
+class UserInvitationSerializer(serializers.ModelSerializer):
+    """
+    Serializer to store the invitation details
+    """
+
+    challenge_title = serializers.SerializerMethodField()
+    challenge_host_team_name = serializers.SerializerMethodField()
+    user_details = serializers.SerializerMethodField()
+
+    class Meta:
+        model = UserInvitation
+        fields = (
+            "email",
+            "invitation_key",
+            "status",
+            "challenge",
+            "user",
+            "challenge_title",
+            "challenge_host_team_name",
+            "invited_by",
+            "user_details",
+        )
+
+    def get_challenge_title(self, obj):
+        return obj.challenge.title
+
+    def get_challenge_host_team_name(self, obj):
+        return obj.challenge.creator.team_name
+
+    def get_user_details(self, obj):
+        serializer = UserDetailsSerializer(obj.user)
+        return serializer.data
