+++ docker/dev/docker.env
@@ -7,7 +7,7 @@ DEBUG=True
 DJANGO_SETTINGS_MODULE=settings.dev
 DJANGO_SERVER=django
 
-HOSTNAME=http://localhost:8888
+HOSTNAME=localhost:8888
 POSTGRES_NAME=postgres
 POSTGRES_USER=postgres
 POSTGRES_PASSWORD=postgres
