+++ docker/prod/docker_production.env
@@ -22,7 +22,7 @@ EMAIL_HOST_PASSWORD=x
 EMAIL_HOST_USER=x
 EMAIL_PORT=123
 EMAIL_USE_TLS=True
-HOSTNAME=https://evalai.cloudcv.org
+HOSTNAME=evalai.cloudcv.org
 MEMCACHED_LOCATION=None
 
 RDS_DB_NAME=x
