+++ frontend/src/views/web/challenge-invite.html
@@ -24,10 +24,10 @@
                                 <div class="row">
                                     <div class="col s12 m12">
                                         <div class="input-field align-left">
-                                            <input type="text" id="username" class="dark-autofill" name="name" ng-model="challenge_invitation.username" ng-minlength="3"
-                                                focus-if="challenge_invitation.username" ng-disabled="challenge_invitation.username" required>
+                                            <input type="text" id="username" class="dark-autofill" name="name" ng-model="challenge_invitation.userDetails.username" ng-minlength="3"
+                                                focus-if="challenge_invitation.userDetails.username" ng-disabled="challenge_invitation.userDetails.username" required>
                                             <span class="form-icon form-icon-dark"><i class="fa fa-user"></i></span>
-                                            <label ng-class="{'active': challenge_invitation.username.length > 0}" for="name">Username*</label>
+                                            <label ng-class="{'active': challenge_invitation.userDetails.username.length > 0}" for="name">Username*</label>
                                             <div class="wrn-msg text-highlight" ng-messages="registerChallengeParticipantForm.username.$error"
                                                 ng-if="registerChallengeParticipantForm.username.$touched || registerChallengeParticipantForm.$submitted">
                                                 <p ng-message="minlength">Username is too short.</p>
